Use multithreading and argparse to streamline downloads

+++ main.py
@@ -1,11 +1,13 @@
 import argparse
+from concurrent.futures import ThreadPoolExecutor
 import os
 import requests
 import pandas as pd
 import pathlib
-import pdb
-from tqdm import tqdm
+import time
 import urllib
+
+#from tqdm import tqdm
 
 
 def parse_book_url(url, title, author, format_="pdf"):
@@ -66,12 +68,12 @@ def book_info_from_table(book_table):
         }
 
 
-def download_book(book, dst_dir, get_epub=False):
+def download_book(book, dst_dir=".", get_epub=False):
     book_dir = dst_dir / book["pkg"]
     book_dir.mkdir(exist_ok=True)
 
     success = {
-        "pdf": False,
+        "pdf": None,
         "epub": None
     }
 
@@ -81,31 +83,46 @@ def download_book(book, dst_dir, get_epub=False):
         request.url, book["title"], book["author"], "pdf"
     )
     pdf_fpath = book_dir / pdf_fname
-    pdf_request = requests.get(pdf_download_url, allow_redirects=True)
 
-    if pdf_status_code == 200:
-        with open(str(pdf_fpath), "wb") as f:
-            ret = f.write(pdf_request.content)
-            success["pdf"] = ret > 0
+    if pdf_fpath.exists():
+        success["pdf"] = "exists"
+    else:
+        success["pdf"] = False
+        pdf_request = requests.get(pdf_download_url, allow_redirects=True)
+        if pdf_request.status_code == 200:
+            with open(str(pdf_fpath), "wb") as f:
+                ret = f.write(pdf_request.content)
+                success["pdf"] = ret > 0
 
     if get_epub:
-        success["epub"] = False
         epub_download_url, epub_fname = parse_book_url(
             request.url, book["title"], book["author"], "epub"
         )
         epub_fpath = book_dir / epub_fname
-        epub_request = requests.get(epub_download_url, allow_redirects=True)
 
-        if epub_request.status_code == 200:
-            with open(str(epub_fpath), "wb") as f:
-                ret = f.write(epub_request.content)
-                success["epub"] = ret > 0
+        if epub_fpath.exists():
+            success["epub"] = "exists"
+        else:
+            success["epub"] = False
+            epub_request = requests.get(epub_download_url, allow_redirects=True)
+            if epub_request.status_code == 200:
+                with open(str(epub_fpath), "wb") as f:
+                    ret = f.write(epub_request.content)
+                    success["epub"] = ret > 0
     book["success"] = success
+    return True
 
 
 def download_books(dst_dir, book_table, get_epub=False, max_workers=8):
     books = list(book_info_from_table(book_table))
-    download_book(books[0], dst_dir, get_epub)
+
+    with ThreadPoolExecutor(max_workers=max_workers) as executor:
+        futures = [
+            executor.submit(download_book, book, dst_dir=dst_dir, get_epub=get_epub)
+            for book in books
+        ]
+    return books
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -114,6 +131,7 @@ if __name__ == "__main__":
         help="Destination directory"
     )
     parser.add_argument("--epub", action="store_true", help="Grab epubs")
+    parser.add_argument("-w", "--max-workers", type=int, default=8)
     args = vars(parser.parse_args())
 
     book_table_url = "".join(
@@ -127,9 +145,17 @@ if __name__ == "__main__":
     dst_dir.mkdir(exist_ok=True)
 
     table_path = dst_dir / "table.xlsx"
-    book_table = pd.read_excel(book_table_url)
 
-    # Save table.
-    book_table.to_excel(str(table_path))
+    if not table_path.exists():
+        book_table = pd.read_excel(book_table_url)
+        book_table.to_excel(str(table_path))
+    else:
+        book_table = pd.read_excel(str(table_path))
 
-    download_books(dst_dir, book_table, get_epub=args["epub"])
+    start_time = time.time()
+    res = download_books(
+        dst_dir, book_table, get_epub=args["epub"],
+        max_workers=args["max_workers"]
+    )
+    elapsed = time.time() - start_time
+    print(elapsed)
